Move dialog components to new package
