+++ app/src/main/java/com/dubproductions/bracket/ui/components/dialogs/ReusableDialog.kt
@@ -1,4 +1,4 @@
-package com.dubproductions.bracket.ui.components
+package com.dubproductions.bracket.ui.components.dialogs
 
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Icon
