+++ app/src/main/java/com/dubproductions/bracket/ui/main/hosting/EditTournament.kt
@@ -28,7 +28,7 @@ import androidx.compose.ui.unit.sp
 import com.dubproductions.bracket.R
 import com.dubproductions.bracket.data.Tournament
 import com.dubproductions.bracket.data.status.TournamentStatus
-import com.dubproductions.bracket.ui.components.ReusableDialog
+import com.dubproductions.bracket.ui.components.dialogs.ReusableDialog
 import com.dubproductions.bracket.data.ui_state.EditTournamentUIState
 import java.text.SimpleDateFormat
 import java.util.Locale
